# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Swagger Example Book API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "http://localhost:8080",
    "paths": {
        "/api/v1/book": {
            "post": {
                "tags": [
                    "book"
                ],
                "summary": "Bookを追加する",
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
        "/api/v1/book/{id}": {
            "put": {
                "tags": [
                    "book"
                ],
                "summary": "Bookを更新する",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Book Id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            },
            "delete": {
                "tags": [
                    "book"
                ],
                "summary": "Bookの削除",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Book Id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
        "/api/v1/books": {
            "get": {
                "tags": [
                    "book"
                ],
                "summary": "リストを返す",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/definitions/model.Book"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "model.Book": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string",
                    "example": "本の内容"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "title": {
                    "type": "string",
                    "example": "本のタイトル"
                }
            }
        }
    }
}